opengl: draw text with minimum texture switching
In GL we need to draw characters into a bitmap and upload that to
the GPU. This is expensive so we cache characters in texture
atlases. Until now when we have drawn from a texture atlas we
have always done this sequentially for every character and changed
textures atlases quite extensively. This is quite bad for
performance. This change deferrs drawing as long as possible and
sorts drawing commands per texture so when we need to draw we
draw with minimum texture rebinding.

Change-Id: I237beaa9f3c3a4bcbd9750e107f017911c967704