opengl: track state of active and bound textures in context
Add TextureState which is responsible to track the state of
texture binding (and unbinding) and changing the current active
texture unit. This is necessary because all GL calls reset the
internal state without checking what the current state actually
is and this can accumulate to a large amount of overhead.

We also unbound the textures (glBindTexture with 0 as ID) after
every operation which is also a major unneeded overhead which is
fixed by this commit.

Change-Id: I770a6a744c0c41850c576b928f027375962088aa