opengl: sync scissor and stencil state, generic capability state
Scissor and stencil test needed to be disabled in flush() (which
means every postDraw call) because sometimes the state became out
of sync with the current state. This commit adds sync() function
which synchronises the actual OpenGL state and adds debugging
mechanisms to warn when the state becomes out of sync (so we can
inspect the exact moment in apitrace).

Added a GenericCapabilityState for GL capabilities like
GL_SCISSORS_TEST, GL_STENCIL_TEST, GL_BLEND,... and refactored
existing ScissorState and StencilState to inherit from it.

Change-Id: Ifc159108a5ce850c78a89b1f5b8d12ecdd84f459