On Windows, compile OpenGL shaders in advance and disable on failure
This is a combination of multiple commits:

Don't claim to be reading a file before trying to open it

Avoid redundancy: The same information was logged in readProgramBinary()

Use VCL_GL_INFO here

Failing to open a cached shader binary is a no reason for a
SAL_WARN. It is normal that they don't exist when first trying, that
is the very nature of a 'cache'.

Move exithelper.h to include

We will want to use EXITHELPER_CRASH_WITH_RESTART in vcl, too.

If the shader compilation or loading of an already compiled shader
fails, disable OpenGL in the registry and exit with the
EXITHELPER_CRASH_WITH_RESTART status. The wrapper process will thus
run soffice.bin once more, and this time OpenGL will not be used.

Change-Id: I3cc4a615f00a8a1adb584493861e4956c83cec32