tdf#105288 MM wait until all emails are send
Seems mail merge is now much faster then the sending of the merged
documents via email. This is why the attached workaround patch to
the tdf#103919 bug report, which simply sleeps after queuing a new
message, helped solving the problem. It slowes down the mail merge
and gives time to the mail dispatcher thread.

But we actually want to wait until all mails are send, so we have
to add a listener and wait for the mail thread to become idle.

This listener also allows us to cancel further mail merge when an
SMTP error has occured.

Change-Id: I9d13bcd8f0d0ff084b20d72ab96f70afa3a764ba