WIN run main thread redirects ignoring SolarMutex
This way we can drop all the special nReleased handling. Instead we use
the same mechanism as on Mac, where we keep the lock, but disable it for
the main thread. As a security measure we assert on duplicate redirects,
which should not happen.

As a result we can't use SendMessage on the main thread itself, which
would normally just call the WinProc directly. This could be accomplished
by converting the redirect bool into a counter, which should be safe, as
no other thread could acquire the SolarMutex, as we don't release it.

Change-Id: Icd87b3da37a2489f3cad2bc80215bf93fc41d388