OSX revert fix for i#90083
Part of commit d6f7c94e5c27ba02ff5c3229760c9808cc9b5bea

At least on my current OSX box application based window switching
"just works" "out of the box", even without the code.

With the replacement of the list with a set, the result is a little
bit unexpected and as all the code has a comment "FIXME: lousy
workaround", my fix is simply to drop it.

Change-Id: I456503a74d8cddbd1e81b4a826b94381b424c78c